requirements.txt
anthropic>=0.40.0
openai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.12

# RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0
markdown>=3.7
beautifulsoup4>=4.12.0

# Export
fpdf2>=2.7.0

# TTS
elevenlabs>=1.0.0

# Telegram
python-telegram-bot>=21.0
